refactor: colocate Hermes input Zod schemas under schemas/
Move article-analysis input schema to src/schemas/ and hermesTickerIdSchema to agent-runtime src/schemas/. Re-export paths unchanged for consumers of @workspace/agent-runtime.

Document content-generation env import: agents must use @mediapulse/env/agents-* subpaths (not root env or process.env).
