+++ packages/shared/agent-runtime/src/schemas/hermes-ticker-id.test.ts
